Ingress: Move to namespaced ingress controllers and edge VIP
This PS moves to namespace specific ingress controllers, and a
cluster wide ingress controller that acts as the external front
door to the cluster. The cluster ingress controller will also
be able to create a dummy interface that can be announced via the
OpenStack-Helm-Infra Calico chart.

The ability to exercise the edge VIP feature is currently only
demonstrated in the single node gate and guide, as it requires
additional configuration, or hardware to set up the routes in
a multinode env.

Change-Id: I8b823b93465f2e90aaabcca9ec9b783d34539e07