TMP FIX: API server to come up on k8s network
For ingress to work api server should come up on k8s network, which is usually
the mgmt network. So this is a temp fix for issue where config-api server was
comin up on control data network.

This will be fixed later once control_data_net_list is fully implemented

Change-Id: I0457a5b42b9633cd3b9ded4c9846429ef1430161
Closes-Bug: #1760694